undefined references for some of the objects, not sure why yet
